src/backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
bcrypt==4.2.1
cryptography==44.0.0
pyjwt==2.10.1
mysql-connector-python==8.3.0
python-dotenv==1.0.0
requests==2.32.3
